Moved tests to suggested location
